This PR brings a massive performance improvement when importing link-check. These are the results on my linux machine:

Before:

```sh
$ time for i in {1..30}; do node -e 'require("./index.js")'; done

real   0m5,393s
user   0m5,424s
sys    0m1,395s
```

After:

```sh
$ time for i in {1..30}; do node -e 'require("./index.js")'; done

real   0m1,678s
user   0m1,032s
sys    0m0,670s
```